Change spoof CRC specification to match ruse.
Previously spoof would use degree reflect polynomial, where the
polynomial is reflected if reflect is true. Now spoof uses degree
polymomial reflect, where the polynomial is never reflected. This
is the same convention that ruse uses. Use the word "degree"
instead of "dimension" for the number of bits in the CRC, also
to be consistent with ruse.